requirements.txt
jieba
torchtext
